init git repository
Initial PINocchio isn't functional yet, internal logic is still being
coded. As now:

- It expects a multithreaded application (not a multiproccess one), as
communication used is based on locks and shared memory.
- Thread holder/syncer is working on static. As now, since sockets are
not being used, 300~500 delays are required to correctly flush on
pthread_no_conflilct.c. It worth noting that I couldn't find a reason
to not use 0 instructions as delay.
- No sleep strategy or function hijack were done, so, as now, it stay
on a kind of a busy wait expecting other threads. It work, nonetheless.
- makefile.rules still a little messy and should receive a clean up.
- No log of any kind is created as now, only stdout is used.