Remove holder and move to THREAD_INFO
Remove every holder, which shouldn't be used anymore, assuming we are not
sending empty work anymore. Remove prints from the end of tool, and change how
controller define if the execution should continue or not (it was depending
entirely on holder).

It worth commenting, applying this patch will remove the notion of delay, used
before to control threads. As now, everyone should be considered unlock unless
lockhash say otherwise.