Include MSG_FINI and correlated
Include MSG_FINI and structures to deal with the thread end. Include check for
the end of all threads, exiting from controller once all threads are gone.

It's important to release a thread ending, since it doesn't exit exactly at the
same instruction it call the thread exit pin hook. A variable to give more than
the needed room is, at least for now, provided on controller.h