Add log to execution
Finally include a functional log to keep track of thread states on a given
execution. Output is on a custom json format to be interpreted by a script or
an external tool not yet provided.

Dump and clean is done when all threads exit, and compilation is included on a
separate shared object in makefile.rules. It merge entries when log become very
large, dealing with multiple states. It uses ofstream to save (instead of
fprintf, as first considered).