Initial tests added
Add a simple but useful pair of python scripts that identifies current examples
and already assert they are running using 1,2,4,8,16 and 32 threads with or
without PINocchio.

It tests for path existence of needed programs and show the results in the end.
Trying to mute or redirect stdout/stderr could be problematic, and is being
avoid at the moment. Using PIPE causes the program to halts, leading to false
negative tests.

To run, just issue: python tests/examples.py