Semaphores: Hijack and create actions
Create actions on sync and treat each of the possible semaphores cases,
following the standard on mutex side. Small fix on lockhash side -
mainly return the result from semaphore_wait as done on try_wait.

Already complete the hijack functions on pinnochio and try to replace
them on the executing code, same as mutex part.