Add exec_tracker initial implementation
Include a new file to track who are the threads running and waiting at a
given moment. It's basically using two heaps:

waiting: min-heap
running: max-heap

The main reason is because we want to get the "smallest" (less
instructions) thread to decide if should awake it or not but, for the
running, we want the max, to see if it will allow the thread to awake.