Exec_tracker: use ordered list instead of heap
Using two heaps was an overkill and wrong, updating threads was causing
a lot of problems in some situations. Use a simpler list, what seems to
be more efficient and better suited for the job.

Keep the start and the end of the list, rewrite exec_tracker functions
and update thread usage.