Deal with watcher_id messing the #1 tid
Somewhat surprising, internal threads have tid as well. Not good for us,
since we index by tid and expose that to the user. Have special handlers
for that but keep the same on sync logic since it's an critical part.

Since we are adding to a global code with easy access, rename error to
log and just have it there.