Add time-based simulation (no PRAM)
Include an option (used by knob, argument to a pin tool) to not simulate
a PRAM but, instead, just keep track of the threads, using time and not
cycles count.

The idea here is: how would a workload behave if there is no control,
without a pram execution. It's basically all our pieces with two
differences:

- There is actually no reason to use exec_tracker. It doesn't get into
the way, though.
- ACTION_DONE is not required, don't add the callback.
- On a trace_bank update, one should know the start (found at start) and
calculate the difference to find the next time.