simple_sempahore_app: fix unbalanced work
The main idea for simple_semaphore_app was to have the same amount of
work divided between all the worker threads. Not only work could end up
not being equally divided, as the final work was changing, since the
computation was changing, with the number of threads.

Fix it to have a more robust simple_semaphore_app example and allow
examples during Benchmark to run for longer.

Update all images from the examples.